requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
tensorflow-hub==0.14.0
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
opencv-python==4.8.0.74